Project structure cleanup + MSVC build fixes

+++ include/memoria/core/packed/array/packed_array_iterator.hpp
@@ -70,7 +70,7 @@ private:
         pos_ -= 1;
     }
 
-    void advance(ssize_t n) {
+    void advance(int64_t n) {
         pos_ += n;
     }
